Bit of work on unit test
